scale flow quality correctly

+++ inc/flow.h
@@ -39,7 +39,7 @@
 /**
  * @brief Computes pixel flow from image1 to image2
  */
-uint32_t compute_flow(uint8_t *image1, uint8_t *image2, float x_rate, float y_rate, float z_rate,
+uint8_t compute_flow(uint8_t *image1, uint8_t *image2, float x_rate, float y_rate, float z_rate,
 		float *histflowx, float *histflowy);
 
 #endif /* FLOW_H_ */
